refactor: added new Validator with Yup and refactored Owner to use the validator
Added Validation class to hold the errors from the entities. Added Validator to validate the
entities, it uses the lib Yup to validate the entity. Refactored Owner core to accomodate and use
both the validation and the validator, to make this change it was necessary to abstract an Entity
class to have the validation and also the ID that is common through the entities. WARNING: Unit
Tests are still not passing because we need to add the validation class to those also.